tests/wpt/tests/web-bundle/subresource-loading/accept-header.https.tentative.html

<!DOCTYPE html>
<title>Accept: request header in webbundle requests</title>
<link
  rel="help"
  href="https://github.com/WICG/webpackage/blob/main/explainers/subresource-loading.md"
/>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/test-helpers.js"></script>
<body>
  <script>
    setup(() => {
      assert_true(HTMLScriptElement.supports("webbundle"));
    });
    promise_test(async () => {
      await addWebBundleElementAndWaitForLoad(
        "../resources/check-accept-header-and-return-bundle.py",
        /*resources=*/ []
      );
    }, "Accept: header in a request for a bundle should contain application/webbundle MIME type.");
  </script>
</body>
